Refactor SourceUtils
